Lint and move tests
