+++ blockchain/state_manager/src/lib.rs
@@ -30,10 +30,9 @@ where
     where
         D: DeserializeOwned,
     {
-        let actor = self.get_actor(addr, ts)?.ok_or(Error::State(format!(
-            "Actor for address: {} does not exist",
-            addr
-        )))?;
+        let actor = self
+            .get_actor(addr, ts)?
+            .ok_or_else(|| Error::State(format!("Actor for address: {} does not exist", addr)))?;
         let act: D = self.bs.get(&actor.state)?.ok_or_else(|| {
             Error::State("Could not retrieve actor state from IPLD store".to_owned())
         })?;
@@ -54,7 +53,7 @@ where
     }
     pub fn get_actor(&self, addr: &Address, ts: &Tipset) -> Result<Option<ActorState>, Error> {
         let state = HamtStateTree::new_from_root(self.bs.as_ref(), ts.parent_state())
-            .map_err(|e| Error::State(e))?;
-        state.get_actor(addr).map_err(|e| Error::State(e))
+            .map_err(Error::State)?;
+        state.get_actor(addr).map_err(Error::State)
     }
 }
